Change rules to be just a redirect
