--- src/app/config/rules/bs_golden.json
@@ -1,82 +1,0 @@
-{
-	"title": "BeeStation Rules",
-	"description": "",
-	"rules": [
-		{	
-			"title": "Admin Discretion",
-			"clauses": [
-				"Admins have the authority to 'ignore' rules when they feel it's in the best interest of the server/playerbase. They will, however, be fully accountable for their actions.",
-				"Listen to them without argument during the round, but if you feel there is a problem with their actions, report them on the forums afterwards. The Admin Conduct can be found here: https://wiki.beestation13.com/view/Admin_Conduct"
-			]
-		},
-		{
-			"title": "Be Excellent to Each Other",
-			"clauses": [
-				"Be patient with other players and assist them when you can. Especially new ones.",
-				"Do not excessively bully/harass other players in or out of game.",
-				"No erotic content is to be posted OOC or IC as the server contains minors. This rule is a zero-tolerance rule and will be met with an unappealable permanent ban from both servers. We may or may not share the logs so others can laugh at you. Go to a containment server if you want to ERP.",
-				"Lawyering rules/silicon laws provided in game is not encouraged.",
-				"Don’t attack/damage the arrivals shuttle, and do not attack anyone until they have left the area at least once.",
-				"Racial and homophobic slurs are not permitted."
-			]
-		},
-		{
-			"title": "Validity Rules",
-			"clauses": [
-				"Once a player is valid, they can be killed or punished by other ingame means.",
-				"Antagonists are valid once proven. Suspicions, hearsay over comms, and meta knowledge are not sufficient enough to deem someone valid, but allow for further investigation.",
-				"Players intentionally cooperating with antagonists are also valid. Self-antags are also valid (act like an antag, be treated like an antag.)",
-				"Admins can deem a player valid.",
-				"Mimes breaking their vow of silence are valid.",
-				"Spamming radio/communications makes you valid.",
-				"Validhunting (seeking/baiting players to kill in validity) in turn, makes you valid.",
-				"Heads of Command/Security cannot make players valid, only warrant their arrest."
-			]
-		},
-		{
-			"title": "No Self-Antagging or Grief",
-			"clauses": [
-				"Self-antagging includes damaging the station, attacking other players without a valid reason, and subverting the established chain of command as a non-antagonist role.",
-				"Griefing includes intentionally ruining the game by abusing ingame mechanics (exploits, force crashes/lag, spamming shuttle delay, ect.)",
-				"You may defend yourself against others in any way you are capable, but you must avoid escalating the situation if possible."
-			]
-		},
-		{
-			"title": "Antag Conduct",
-			"clauses": [
-				"Antagonists are expected to follow any rules that do not interfere with their objective.",
-				"Otherwise antagonists are free to complete their objective by any means available.",
-				"You must work with fellow antags towards completing a group objective if there is one.",
-				"Traitors, Blood Brothers, and Changelings must complete half of their objectives before engaging in mass murder/destruction."
-			]
-		},
-		{
-			"title": "Ahelp and Ticket Conduct",
-			"clauses": [
-				"Leaving/Ignoring an admin PM will automatically label you guilty.",
-				"Showing respect and patience with admins has proven to make them more lenient.",
-				"An admins ruling on a issue is final for that round. Do not try to bring it up again.",
-				"Do not immediately ask if the person who killed you is an antag, you will be rejected. If you are still uncertain if your death was valid, spectate them and see if they are performing additional unruly behavior before reporting."
-			]
-		},
-		{
-			"title": "Meta and Communications",
-			"clauses": [
-				"Using outside sources of information/software on the round to gain an advantage over players is not tolerated.",
-				"In character communication in out of character channels (ICK OCK) is not allowed. Discussing out-of-round information ingame is also not tolerated. This refers to admin tickets and references to reporting other players (eg. \"I'm ahelping you for doing that\") as well as any information that was obtained through channels considered as ICK OCK.",
-				"Recordings and Streaming are allowed. Streams are required to have a minimum 1 minute delay for broadcasting and are responsible for any metacommunications in their chat feed. Recordings must not be uploaded until the round has ended.",
-				"You cannot attack players over events that occured in a previous life. This includes ghost respawns, and new rounds. This does not include cloning.",
-				"If you wish to use metacomms to assist new players, that is fine. Please contact an admin before doing so."
-			]
-		},
-		{
-			"title": "Common Sense",
-			"clauses": [
-				"Common sense is expected.",
-				"For specific rulings, follow ingame information or those specified on our forums <a href=\"https://forums.beestation13.com/t/council-rulings/42\">here</a>.",
-				"You are welcome to ask questions on our <a href=\"https://discord.gg/Y7RTtkK\">Discord</a>.",
-				"Have fun!"
-			]
-		}
-	]
-}
