+++ src/app/config/servers.yml
@@ -11,6 +11,7 @@
   port: 7777
   open: true
   color: '0, 150, 0'
+  rules_url: 'https://wiki.beestation13.com/view/Rules'
 
 - id: 'bs_sage'
   aliases:
@@ -25,6 +26,7 @@
   port: 7878
   open: true
   color: '0, 0, 150'
+  rules_url: 'https://wiki.beestation13.com/view/Rules'
 
 - id: 'nsv_tupelo'
   aliases:
@@ -37,6 +39,7 @@
   port: 1337
   open: true
   color: '150, 0, 0'
+  rules_url: 'https://nsv.beestation13.com/wiki/Rules'
 
 - id: 'bs_clover'
   aliases:
@@ -50,3 +53,5 @@
   port: 8888
   open: true
   color: '84, 123, 16'
+  rules_url: 'https://wiki.beestation13.com/view/Rules'
+
