--- src/app/templates/rules.html
@@ -1,33 +1,0 @@
-{% extends "layout.html" %}
-{% block title %}Rules - Roleplay{% endblock %}
-{% block active %}rules{% endblock %}
-
-{% block head %}
-	<meta property="og:title" content="BeeStation - {{ rules["title"] }}" />
-{% endblock %}
-
-{% block content %}
-
-	<center>
-
-	<div style="text-align:center;flex-basis:33.33%;"><a href="/rules" class="button">LRP</a><a> </a><a href="/rules/rp" class="button">MRP</a></div>
-
-
-	<h1>{{ rules["title"] }}</h1>
-	<span class="rule-desc" style="color: var(--bee-gray);">{{ rules["description"] }}<span>
-	</center>
-
-	<br><br>
-
-	{% for rule in rules["rules"] %}
-		<div class="rule">
-		<span class="rule-title">{{ loop.index }}. {{ rule["title"] }}</span><br>
-		<span class="rule-desc">
-		{% for clause in rule["clauses"] %}
-			<br><br>{{ clause|safe }}
-		{% endfor %}
-		</span>
-		</div>
-	{% endfor %}
-
-{% endblock %}
